techpath-project/requirements.txt
Django
djangorestframework
django-cors-headers
gunicorn
whitenoise # Added for static files, good practice with Django
